BugID:21767265: Change vfs_lseek the parameter type

+++ 3rdparty/experimental/fs/uffs/uffs_vfs.c
@@ -283,10 +283,10 @@ static int32_t uffs_vfs_write(vfs_file_t *fp, const char *buf, uint32_t len)
     return nbytes;
 }
 
-static uint32_t uffs_vfs_lseek(vfs_file_t *fp, uint32_t off, int32_t whence)
+static uint32_t uffs_vfs_lseek(vfs_file_t *fp, int64_t off, int32_t whence)
 {
-    uint32_t ret;
-    int32_t  fd;
+    int64_t ret;
+    int32_t fd;
 
     fd = (int32_t)(fp->f_arg);
     ret = uffs_seek(fd, off, whence);
